add utilities
for progress, switch out backend for events to redis so multiple workers / hosts
can share the event queue. use contextvars to parse out session id and progress
tokens. use SSE to report back progress if a stream is open for a session that
corresponds to the session in a client request that included a progress token.